requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
playwright==1.40.0
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
lxml==4.9.3
openpyxl==3.1.2
plotly==5.18.0
scipy==1.11.4
scikit-learn==1.3.2
networkx==3.2.1
geopandas==0.14.1
